Automate workaround for etcd regression
In 4.4, single node installs will fail because of a change in etcd which
started to require clusters with at least 3 master nodes.
This can be worked around by setting an override on the etcd cluster
resource:
https://github.com/code-ready/snc/wiki/Workaround-for-4.4--etcd-operator-addition
Since this is not automated yet in the installer, we can do it in snc
for now while this is properly fixed in OpenShift.