Add default argument value to createdisk.sh
snc.sh hardcodes the use of crc-tmp-install-data for its data, but
createdisk.sh requires an argument specifying where the install data
is, and errors out if it's missing.
This commit uses `crc-tmp-install-data` as the default value for this
argument if it's missing. If an argument is specified, then it will
override this default value.